+++ src/database/schemas/Guild.js
@@ -1,5 +1,5 @@
 const mongoose = require("mongoose");
-const { CACHE_SIZE, PREFIX_COMMANDS } = require("@root/config.js");
+const { CACHE_SIZE, PREFIX_COMMANDS, STATS } = require("@root/config.js");
 const FixedSizeMap = require("fixedsize-map");
 const { getUser } = require("./User");
 
@@ -17,6 +17,13 @@ const Schema = new mongoose.Schema({
     bots: { type: Number, default: 0 },
   },
   prefix: { type: String, default: PREFIX_COMMANDS.DEFAULT_PREFIX },
+   stats: {
+    enabled: Boolean,
+    xp: {
+      message: { type: String, default: STATS.DEFAULT_LVL_UP_MSG },
+      channel: String,
+    },
+  },
   ticket: {
     log_channel: String,
     limit: { type: Number, default: 10 },